refactor: Extract clusters_from_pairs into shared _benchmark_utils
Amazon-Google's connected-components/singleton-completion helper is
now needed verbatim by the new Abt-Buy loader too -- move it to
_benchmark_utils.clusters_from_pairs (generic, already the home for
every other genuinely-shared benchmark primitive) and make
amazon_google._clusters_from_pairs a thin wrapper so existing call
sites/tests are unaffected. No behavior change (29/29 existing
Amazon-Google tests pass unmodified).

Also vendors the Abt-Buy CSVs (matchbench/Abt-Buy on Hugging Face)
for the upcoming loader.
